Make interrupt() implementation safer
interrupt() only has effects if the thread object/id refers to an actual
thread. This behavior is hinted by pthread_cancel()'s ESRCH error code
(check the manpage). Another hint can be found in Boost.Thread
documentation as listing:

"If *this refers to a thread of execution"
  -- Member function interrupt() EXTENSION

"the boost::thread object ceases to represent the now-detached thread"
  -- Detaching thread

Other references can be found as this being an expected behavior and
common practice.

As for our implementation, this change prevents the use of the executor
(pimpl_->executor.post()) after pimpl_->executor.on_work_finished() has
already been called (done in fiber.join() and fiber.detach()).

The safety is useful in the scoped_thread idiom presented in
Boost.Thread:

    struct interrupt_and_join_if_joinable
    {
      template <class Thread>
      void operator()(Thread& t)
      {
        t.interrupt();
        if (t.joinable())
        {
          t.join();
        }
      }
    };

No “error” information needs to be propagated as (1) the user can
already extract such info through other ways, (2) the abstraction
depends on user code structure and not on external events, and (3)
it fits well with design of other libraries in the wild.