Remove semaphore
A semaphore that behaves adequately to the user scenario just happens to
need too many policies (what to do on overflow, can the notifier yield
to another fiber, should the notifier suspend, how can the waiter
consume several events safely avoiding the value's overflow...).

It's time to ditch it. Should the user need a semaphore, he can just
implement it easily on top of a condvar.